Interface Adjustments CashDesk & implementation of adapters

+++ Domain/CashDesk/IBankService.cs
@@ -2,7 +2,7 @@ namespace Domain.CashDesk;
 
 public interface IBankService
 {
-    Task<BankTransactionContext> CreateTransactionContextAsync(double amount);
+    Task<BankTransactionContext> CreateTransactionContextAsync(int amount);
     
     Task<AuthorizationResult> AuthorizePaymentAsync(BankTransactionContext context, string token);
